remove extra files
